refactor: refactored the automatic_push directory
